debugged and tested new Simulation_Parameter object and its corresponding structures and actions

+++ capitalism/economy/actions/initialize.py
@@ -1,8 +1,8 @@
-from economy.models.states import Project, TimeStamp
+from economy.models.states import Project, TimeStamp, Simulation_Parameter
 from economy.models.report import Trace
 from economy.models.commodity import Commodity
-from economy.models.owners import Industry, SocialClass
-from economy.models.stocks import IndustryStock, SocialStock
+from economy.models.owners import Industry, SocialClass, StockOwner
+from economy.models.stocks import IndustryStock, SocialStock, Stock
 from economy.actions.exchange import set_initial_capital, set_total_value_and_price
 from django.contrib.staticfiles.storage import staticfiles_storage
 import pandas as pd
@@ -18,14 +18,16 @@ from django.urls import reverse
 #! If this should happen, the simulations will still exist and function, but the user may lose track of which is which
 #! Because the descriptions will either not be accessible or may have changed
 #! TODO make this more foolproof
-
 def initialize_projects(request):
     logged_in_user=request.user
     logger.info(f"Initialise project table {logged_in_user}")
-    #! TODO test that this is the admin user
+    if not logged_in_user.is_superuser:
+        logger.warning(f"User {logged_in_user} wanted to rebuild the project table and was prevented from doing so")
+        messages.warning("Sorry, only the administrator can carry out this function")
+        return HttpResponseRedirect(reverse("economy"))
     try:
         file_name = staticfiles_storage.path('data/projects.csv')    
-        logger.info (f"Initializing projects from file {file_name} for user {logged_in_user}")
+        logger.info (f"User {logged_in_user} is initializing the project table from file {file_name}")
         Project.objects.all().delete()
         df = pd.read_csv(file_name)
         for row in df.itertuples(index=False, name='Pandas'):
@@ -37,29 +39,37 @@ def initialize_projects(request):
         logger.error(f"Could not load projects because of {error}")
         messages.error(request, f"Sorry, could not load the project file because of {error}")
         return HttpResponseRedirect(reverse("economy"))
-     
-#! Creates a fresh, complete dataset for a single user
+
+
+#! Create a fresh, complete dataset for a single user
 #! Wipes out anything they already have so they can start a new simulation
 #! TODO restart a single simulation should be a separate activity
 #! TODO and this should probably also be an admin action and also automatically performed on signup
-
 def initialize(request):
     logged_in_user=request.user
     logger.info(f"User {logged_in_user} has asked to reinitialize")
     #TODO 'Do you really want to do this?'
     #TODO option to re-initialize just one project 
-    #! Timestamps    
+
+#! Timestamps    
     #! TODO insist that there is only one time-stamp per project in the CSV file
     Trace.objects.filter(user=logged_in_user).delete()
     Trace.enter(request.user,0, "INITIALIZE ENTIRE SIMULATION")
+    Simulation_Parameter.objects.filter(user=logged_in_user).delete()
+    Commodity.objects.filter(user=logged_in_user).delete()
+    Stock.objects.filter(user=logged_in_user).delete()
+    StockOwner.objects.filter(user=logged_in_user).delete()
     TimeStamp.objects.filter(user=logged_in_user).delete()
+
     file_name = staticfiles_storage.path('data/timestamps.csv')    
     logger.info( f"Reading time stamps from {file_name}")
     df = pd.read_csv(file_name)
 
     for row in df.itertuples(index=False, name='Pandas'):
+        s=Simulation_Parameter(user=request.user, project_number=row.project_FK) #! Default everything else
+        s.save()       
         time_stamp = TimeStamp(
-            project_number=row.project_FK,
+            simulation_FK=s,
             period=row.period,
             step=row.description,
             stage="M_C", #! projects must start with this stage TODO remove this field from the CSV file
@@ -67,24 +77,20 @@ def initialize(request):
             user=logged_in_user,
         )
         time_stamp.save()
-        time_stamp.comparator_time_stamp_FK=time_stamp #! first stamp has no navel
+        time_stamp.comparator_time_stamp_FK=time_stamp #! comparator for the first stamp is itself
         time_stamp.save()
-    #! Start off with project 1
-    time_one=TimeStamp.objects.get(user=logged_in_user,project_number=1)
-    logged_in_user.current_time_stamp=time_one
-    logged_in_user.save()
 
-    #! Basic setup complete, now read the data files
-    #! Commodities
+#! Commodities
     Commodity.objects.filter(user=logged_in_user).delete()
     file_name = staticfiles_storage.path('data/commodities.csv')    
     logger.info( f"Reading commodities from {file_name}")
     df = pd.read_csv(file_name)
     for row in df.itertuples(index=False, name='Pandas'):
-        this_time_stamp=TimeStamp.objects.get(project_number=row.project,user=logged_in_user)        
-        logger.info(f"Creating commodity {(row.name)} for user {logged_in_user} and project {row.project} with time stamp {this_time_stamp}")
+        logger.info(f"Creating commodity {(row.name)} for user {logged_in_user} and project {row.project} ")
+        simulation=Simulation_Parameter.objects.get(project_number=row.project,user=logged_in_user)
+        time_stamp=TimeStamp.objects.get(simulation_FK=simulation,user=logged_in_user)
         commodity = Commodity(
-            time_stamp_FK=this_time_stamp,
+            time_stamp_FK=time_stamp,
             name=row.name,
             origin=row.origin_type,
             unit_value=row.unit_value,
@@ -101,19 +107,20 @@ def initialize(request):
         )
         commodity.save()
     
-    #!Industries
+#!Industries
     Industry.objects.filter(user=logged_in_user).delete()
     file_name = staticfiles_storage.path('data/industries.csv')    
     logger.info( f"Reading industries from {file_name}")
     df = pd.read_csv(file_name)
     for row in df.itertuples(index=False, name='Pandas'):
-        this_time_stamp = TimeStamp.objects.get(project_number=row.project, user=logged_in_user)
-        this_commodity=Commodity.objects.get(time_stamp_FK=this_time_stamp, name=row.commodity_name, user=logged_in_user)
-        logger.info(f"Creating Industry {(row.industry_name)} with output {this_commodity} for user {logged_in_user} and project {row.project} with time stamp {this_time_stamp}")
+        simulation=Simulation_Parameter.objects.get(project_number=row.project,user=logged_in_user)
+        time_stamp=TimeStamp.objects.get(simulation_FK=simulation, user=logged_in_user)
+        commodity=Commodity.objects.get(time_stamp_FK=time_stamp, name=row.commodity_name, user=logged_in_user)
+        logger.info(f"Creating Industry {(row.industry_name)} with output {commodity} for user {logged_in_user} and project {row.project} with time stamp {time_stamp} in simulation {simulation}")
         industry = Industry(
-            time_stamp_FK=this_time_stamp,
+            time_stamp_FK=time_stamp,
             name=row.industry_name,
-            commodity_FK=this_commodity,
+            commodity_FK=commodity,
             output_scale=row.output,
             output_growth_rate=row.growth_rate,
             current_capital=0,
@@ -124,19 +131,19 @@ def initialize(request):
         )
         industry.save()
 
-    #! Social Classes
+#! Social Classes
     SocialClass.objects.filter(user=logged_in_user).delete()
     file_name = staticfiles_storage.path('data/social_classes.csv')
     logger.info( f"Reading social classes from {file_name}")
     df = pd.read_csv(file_name)
     for row in df.itertuples(index=False, name='Pandas'):
-        this_time_stamp = TimeStamp.objects.get(project_number=row.project, user=logged_in_user)
-        # labour_power=Commodity.objects.get(time_stamp_FK=this_time_stamp, name="Labour Power", user=logged_in_user),
-        logger.info(f"Creating Social Class {(row.social_class_name)} for user {logged_in_user} and project {row.project} with time stamp {this_time_stamp}")
+        simulation=Simulation_Parameter.objects.get(project_number=row.project,user=logged_in_user)
+        time_stamp=TimeStamp.objects.get(simulation_FK=simulation, user=logged_in_user)
+        logger.info(f"Creating Social Class {(row.social_class_name)} for user {logged_in_user} and project {row.project} with time stamp {time_stamp}")
         social_class = SocialClass(
-            time_stamp_FK=this_time_stamp,
+            time_stamp_FK=time_stamp,
             name=row.social_class_name,
-            commodity_FK=Commodity.objects.get(time_stamp_FK=this_time_stamp, name="Labour Power", user=logged_in_user),
+            commodity_FK=Commodity.objects.get(time_stamp_FK=time_stamp, name="Labour Power", user=logged_in_user),
             stock_owner_type=SOCIAL_CLASS,
             population=row.population,
             participation_ratio=row.participation_ratio,
@@ -146,26 +153,20 @@ def initialize(request):
         )
         social_class.save()
 
-    #! Stocks
-    # ?Can the below be achieved by deleting all Stock objects?
-    # ?And if I delete the children, do the parent objects persist?
-    # ? See https://stackoverflow.com/questions/9439730/django-how-do-you-delete-child-class-object-without-deleting-parent-class-objec
-    # ? Find out by doing it.
-    IndustryStock.objects.filter(user=logged_in_user).delete()
-    SocialStock.objects.filter(user=logged_in_user).delete()
+#! Stocks
     file_name = staticfiles_storage.path('data/stocks.csv')        
     logger.info( f"Reading stocks from {file_name}")
     df = pd.read_csv(file_name)
 
-    # TODO can the below be done more efficiently?
     for row in df.itertuples(index=False, name='Pandas'):
-        this_time_stamp = TimeStamp.objects.get(project_number=row.project, user=logged_in_user)
+        simulation=Simulation_Parameter.objects.get(project_number=row.project,user=logged_in_user)
+        time_stamp=TimeStamp.objects.get(simulation_FK=simulation, user=logged_in_user)
         if row.owner_type == "CLASS":
-            social_class=SocialClass.objects.get(time_stamp_FK=this_time_stamp, name=row.name,user=logged_in_user)
-            commodity=Commodity.objects.get(time_stamp_FK=this_time_stamp, name=row.commodity, user=logged_in_user)
+            social_class=SocialClass.objects.get(time_stamp_FK=time_stamp, name=row.name,user=logged_in_user)
+            commodity=Commodity.objects.get(time_stamp_FK=time_stamp, name=row.commodity, user=logged_in_user)
             logger.info(f"Creating a stock of commodity {(commodity.name)} of usage type {row.stock_type} for class {(social_class.name)} ")
             social_stock = SocialStock(
-                time_stamp_FK=this_time_stamp,
+                time_stamp_FK=time_stamp,
                 social_class_FK=social_class,
                 stock_owner_FK=social_class,
                 commodity_FK=commodity,
@@ -181,11 +182,11 @@ def initialize(request):
             
             social_stock.save()
         elif row.owner_type == "INDUSTRY":
-            industry=Industry.objects.get(time_stamp_FK=this_time_stamp, name=row.name, user=logged_in_user)
-            commodity=Commodity.objects.get(time_stamp_FK=this_time_stamp, name=row.commodity, user=logged_in_user)
+            industry=Industry.objects.get(time_stamp_FK=time_stamp, name=row.name, user=logged_in_user)
+            commodity=Commodity.objects.get(time_stamp_FK=time_stamp, name=row.commodity, user=logged_in_user)
             logger.info(f"Creating a stock of {(commodity.name)} of usage type {row.stock_type} for industry {(industry.name)}")
             industry_stock = IndustryStock(
-                time_stamp_FK=this_time_stamp,
+                time_stamp_FK=time_stamp,
                 industry_FK=industry,
                 stock_owner_FK=industry,
                 commodity_FK=commodity,
@@ -201,6 +202,13 @@ def initialize(request):
             industry_stock.save()
         else:
             logger.error(f"++++UNKNOWN OWNER TYPE++++ {row.owner_type}")
+
+#! set the user up so ta points to project 1
+    simulation=Simulation_Parameter.objects.get(user=logged_in_user, project_number=1)
+    time_stamp=TimeStamp.objects.get(user=logged_in_user,simulation_FK=simulation)
+    logged_in_user.current_time_stamp=time_stamp
+    logged_in_user.save()
+ 
     set_total_value_and_price(user=logged_in_user)
     set_initial_capital(user=logged_in_user)
